requirements.txt
groq==0.11.0
python-dotenv==1.0.0
pandas==2.1.4
schedule==1.2.0